Dices: Added enums for dice names.  Added the D6, which was missing.
Items: Slightly altered IArmor, reworked class Armor to be an abstract class, renamed and reworked armor types that Torbjörn made.  Created an abstract class weapon.

TurnManager: I didn't do anything to this, but because I had to install NUnit locally so that the tests showed up it now says I've altered this.